refactor: refactor forms, separate validation module
